% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctchange.R
\name{pctchange}
\alias{pctchange}
\title{Percentage Change}
\usage{
pctchange(year)
}
\arguments{
\item{year}{The year of bills to be read. (2001-2016)}
}
\value{
Bar charts showing the percentage change of components of school bill in \code{year} and the inflation rate.
}
\description{
Visualizes percentage change of components of school bills data from \code{readbills}.
}

