% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbills.R
\name{plotbills}
\alias{plotbills}
\title{Plot Bills}
\usage{
plotbills(year)
}
\arguments{
\item{year}{The year of bills to be read. (2000-2016)}
}
\value{
Bar charts showing components of school bill in \code{year}.
}
\description{
Visualizes components of school bills data from \code{readbills}.
}

