% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictbills.R
\name{predictbills}
\alias{predictbills}
\title{Predict Bills}
\usage{
predictbills(year)
}
\arguments{
\item{year}{Sets the year of bills to predict}
}
\value{
The cost of school bills at \code{year}
}
\description{
Predicts the cost of school bills at \code{year} based on a linear model
}

