% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictpc.R
\name{predictpc}
\alias{predictpc}
\title{Predict Percentage Change}
\usage{
predictpc(year)
}
\arguments{
\item{year}{Sets the year of bills to predict}
}
\value{
The percentage change of school bills at \code{year}
}
\description{
Predicts the percentage change of school bills at \code{year} based on an exponential model
}

