% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbills.R
\name{readbills}
\alias{readbills}
\title{Read College Bills}
\usage{
readbills(year)
}
\arguments{
\item{year}{Sets the year of bills to be read. (2000-2016)}
}
\value{
A list of fees for the year \code{year}
}
\description{
Scan for the college bills from text file and create a list of fees.
}

