% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinflation.R
\name{readinflation}
\alias{readinflation}
\title{Read Inflation}
\usage{
readinflation()
}
\value{
A list of inflation rates from 2000-2016
}
\description{
Scan the inflation rate from text file and create a list.
}

