% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showstats.R
\name{showstats}
\alias{showstats}
\title{Show Statistics}
\usage{
showstats("type")
}
\arguments{
\item{type}{The kind of analysis to be done. The possible options are \code{summary}, \code{bills}, \code{pctchange}, \code{inflation}, \code{predictbills}, \code{predictpc} and \code{inflationbills}. \code{summary} displays the total amount, percentage increase and total increase of school bills and the inflation rate of each year. \code{bills} displays the timeplot of school bills across the years. \code{pctchange} shows a timeplot of percentage increase of bills across the years. \code{inflation} shows a timeplot of percentage increase of school bills compared to inflation. \code{predictbills} shows a plot of the amount of school bills and the model used to predict future school bills. \code{predictpc} shows a plot of the percentage increase of school bills and the model used to predict future percentage increases. \code{inflationbills} shows a plot of a comparison between the actual cost of school bills and the cost of school bills if it has been increasing at the inflation rate since 2000.}
}
\value{
Results from the kind of analysis specified by \code{type}.
}
\description{
Uses bills data from the read function to perform appropriate analysis and generate graphics associated with the data.
}

